Changed double to float

+++ engine/src/main/java/games/temporalstudio/temporalengine/physics/PhysicsEngineLifeCycle.java
@@ -4,5 +4,5 @@ import games.temporalstudio.temporalengine.LifeCycle;
 import games.temporalstudio.temporalengine.LifeCycleContext;
 
 public interface PhysicsEngineLifeCycle extends LifeCycle {
-	public abstract void compute(LifeCycleContext context, double deltaTime);
+	public abstract void compute(LifeCycleContext context, float deltaTime);
 }
